Introduce CT finder in subdir/finish sorting stage

+++ src/prepare_recon.py
@@ -7,12 +7,41 @@ Script for preprocessing batches of rb82 raw data
 
 @author: michellef
 """
-
+import re
 import os
 import argparse
 from shutil import rmtree
-from pathlib import Path 
-from identify_LM_date import write_pickle, sort_patients
+from pathlib import Path
+from identify_LM_date import sort_patients
+
+
+# Find patients by tag
+def find_patients(dir_path):
+    c = 0
+    ct_paths = []
+    for (dirpath, dirnames, filenames) in os.walk(dir_path): 
+        dirname = str(Path(dirpath).relative_to(dir_path)) 
+        if 'ACCT/' in str(dirname): 
+            new_path = os.path.join(dir_path, dirname)
+            ct_paths.append(new_path)
+            c+=1
+    print(f'{c} moved.')
+    return ct_paths
+
+
+def move_patients(dir_path):
+    ct_paths = find_patients(dir_path)
+
+    for ct in ct_paths:
+        name = (re.search('(\/homes\/michellef\/my_projects\/rb82_data\/rb82_Mar23)\/(?<=\/)(.*)', ct)).group(2)
+        dest = os.path.join('/homes/claes/data_shared/michelle', name)
+        dest = str(dest)
+        ct = str(ct)
+        #print(dest)
+        if not os.path.exists(dest):
+            os.makedirs(dest)
+
+        os.system(f"mv '{ct}' '{dest}'")
 
 
 def remove_patients(data_path, pkl_name):
@@ -24,11 +53,14 @@ def remove_patients(data_path, pkl_name):
     if args.delete:
 
         for t in tracer:
-            rmtree(os.path.join(data_path, t))
+            if os.path.exists(os.path.join(data_path, t)):
+                rmtree(os.path.join(data_path, t))
         for d in dose:
-            rmtree(os.path.join(data_path, d))
+            if os.path.exists(os.path.join(data_path, d)):
+                rmtree(os.path.join(data_path, d))
 
         print(f'Deleted {num_p} patients...')
+
 
 if __name__ == "__main__":
     # Initiate the parser
@@ -37,7 +69,7 @@ if __name__ == "__main__":
     # Add long and short argument
     required_args.add_argument("--data", "-d", help="Data source directory path", required=True)
     parser.add_argument('--delete', dest='delete', type=bool, default=False, help='delete wrong patient types: True/False')
-    required_args.add_argument("--pkl", "-p", help=".pickle file name", required=True)
+    required_args.add_argument("--pkl", "-p", help=".pickle file name")
 
     # Read arguments from the command line
     args = parser.parse_args()
